added sql schema

+++ src/main/java/pl/edu/pw/mini/ingreedio/api/model/UserProducts.java
@@ -1,7 +1,7 @@
 package pl.edu.pw.mini.ingreedio.api.model;
 
 public class UserProducts {
-    private Integer id;
-    private Integer userId;
-    private Integer productId;
+    private Long id;
+    private Long userId;
+    private Long productId;
 }
